splitting up legacy helpers out and removing appInstance
