+++ app/instance-initializers/ember-intl.js
@@ -9,8 +9,6 @@ import { filterBy } from 'ember-intl/utils/initialize';
 
 export function instanceInitializer(instance) {
     let service = instance.container.lookup('service:intl');
-
-    service.set('appInstance', instance);
     service.createLocale = service.createLocale.bind(service, instance);
 
     filterBy(ENV, 'cldrs').forEach((key) => {
